+++ src/components/Form/Input/BaseInput.tsx
@@ -1,9 +1,9 @@
-import { InputHTMLAttributes, ReactNode, useCallback, useState } from 'react'
+import { InputHTMLAttributes, ReactNode, useCallback, useRef, useState } from 'react'
 import CheckCircleIcon from '@heroicons/react/outline/CheckCircleIcon'
 import XCircleIcon from '@heroicons/react/outline/XCircleIcon'
 import InformationCircleIcon from '@heroicons/react/outline/InformationCircleIcon'
 import { composeClasses } from 'lib/classes'
-import { inputVariants, InputVariant as InputVariantType, InputType, getClassesByPseudoClass } from '../shared'
+import { inputVariants, InputVariant as InputVariantType, InputType, getClassesByPseudoClass, getAnimationLabel, getPaddingInput } from '../shared'
 import { Padding, ShadowVariants } from '../../../interfaces/types'
 
 export interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
@@ -57,12 +57,17 @@ function BaseInput({
     large,
     boxShadow = 'lg',
     style,
+    placeholder,
+    value,
     ...otherProps
 }: InputProps) {
     const [focused, setFocused] = useState(false)
     const { disabled } = otherProps
     variant = disabled ? 'disabled' : variant
+    const isDisabled = variant === 'disabled'
     const { input, text } = inputVariants[variant]
+    const inputRef = useRef<HTMLInputElement>(null)
+    const isLabelScalded = !label || focused || value?.toLocaleString().length || inputRef.current?.value.length
 
     const handleFocus = useCallback(
         (event: React.FocusEvent<HTMLInputElement>) => {
@@ -83,17 +88,18 @@ function BaseInput({
     const styles = {
         adornment: composeClasses('text-gray-400 transition duration-500 ease-out focus:ease-in', classNameAdornment),
         container: composeClasses(
-            'placeholder-gray-400 mt-1 flex items-center justify-between bg-transparent font-medium',
+            'gap-3 placeholder-gray-400 mt-1 flex items-center justify-between bg-transparent font-medium',
             'border-solid border',
             'transition duration-500 ease-out focus:ease-in',
-            `hover:shadow-${boxShadow} hover:border-gray-500`,
+            !isDisabled && `hover:shadow-${boxShadow} hover:border-gray-500`,
             className && disabled && getClassesByPseudoClass(className, 'disabled'),
             inputBlank && 'border-none',
             rounded && `rounded-${rounded}`,
             !['error', 'success', 'warning'].includes(variant) && focused && 'border-blue-500',
-            variant === 'disabled' && 'bg-gray-100 text-gray-400 cursor-not-allowed',
+            ['error', 'success', 'warning'].includes(variant) ? 'bg-white' : 'bg-gray-50',
+            isDisabled ? 'bg-gray-100 text-gray-400 cursor-not-allowed' : 'hover:bg-white',
             input.borderColor,
-            padding,
+            padding && `p-${padding}`,
             !padding && paddingX && `px-${paddingX}`,
             !padding && paddingY && `py-${paddingY}`,
             input.color,
@@ -110,22 +116,28 @@ function BaseInput({
                         {startAdornment}
                     </div>
                 )}
-                <div className="flex flex-col mx-3.5 w-full">
+                <div className="flex flex-col w-full relative h-11">
                     {label && (
                         <label
-                            style={{ cursor: 'inherit' }}
-                            className={composeClasses('w-full block text-xxs  font-medium leading-none', variant !== 'disabled' && 'text-gray-500')}
+                            style={getAnimationLabel(!!isLabelScalded)}
+                            className={composeClasses(
+                                'absolute w-full block text-xxs font-medium leading-none text-left whitespace-nowrap overflow-hidden overflow-ellipsis',
+                                !isDisabled && 'text-gray-500'
+                            )}
                         >
                             {label}
                         </label>
                     )}
                     <input
+                        ref={inputRef}
                         {...otherProps}
+                        placeholder={isLabelScalded ? placeholder : ''}
                         className={composeClasses('outline-none w-full font-medium bg-transparent')}
                         onFocus={handleFocus}
                         onBlur={handleBlur}
-                        style={{ cursor: 'inherit' }}
-                        disabled={variant === 'disabled'}
+                        style={{ cursor: 'inherit', zIndex: 1, ...getPaddingInput(!!label) }}
+                        disabled={isDisabled}
+                        value={value}
                     />
                 </div>
                 {endAdornment && (